PR hw-4 methods transferred to another package
