+++ index/index.go
@@ -1,12 +1,7 @@
 package index
 
 import (
-	"encoding/json"
-	"fmt"
-	"io/ioutil"
 	"log"
-	"net/http"
-	"strings"
 	"sync"
 
 	"github.com/polisgo2020/search-Arkronzxc/util"
@@ -15,11 +10,6 @@ import (
 )
 
 type Index map[string][]string
-
-type searchResponse struct {
-	Filename    string `json:"filename"`
-	WordCounter int    `json:"wordCounter"`
-}
 
 // CreateInvertedIndex returns map where key is a word in file, value is filename
 func CreateInvertedIndex(files []string) (*Index, error) {
@@ -66,9 +56,9 @@ func ConcurrentBuildFileMap(wg *sync.WaitGroup, filename string, mapChan chan<- 
 	mapChan <- m
 }
 
-// buildSearchIndex searches by index and returns the structure where the key is the file name, and the value is the
+// BuildSearchIndex searches by index and returns the structure where the key is the file name, and the value is the
 // number of words from the search query that were found in this file
-func (m *Index) buildSearchIndex(searchArgs []string) (map[string]int, error) {
+func (m *Index) BuildSearchIndex(searchArgs []string) (map[string]int, error) {
 	ans := make(map[string]int)
 
 	var cleanData []string
@@ -90,100 +80,4 @@ func (m *Index) buildSearchIndex(searchArgs []string) (map[string]int, error) {
 		}
 	}
 	return ans, nil
-}
-
-func SearchHandler(input string) http.HandlerFunc {
-
-	searchIndex, err := unmarshalFile(input)
-	if err != nil {
-		panic(err)
-	}
-
-	return func(writer http.ResponseWriter, request *http.Request) {
-
-		writer.Header().Set("Content-Type", "application/json")
-
-		parsedSearchPhrase, err, errCode := parseSearchPhrase(request)
-		if err != nil {
-			log.Print(err)
-			http.Error(writer, http.StatusText(errCode), errCode)
-			return
-		}
-
-		resp, err, errCode := answerFormation(searchIndex, parsedSearchPhrase)
-		if err != nil {
-			log.Print(err)
-			http.Error(writer, http.StatusText(errCode), errCode)
-			return
-		}
-
-		finalJson, err := json.Marshal(resp)
-		if err != nil {
-			log.Print(err)
-			http.Error(writer, http.StatusText(http.StatusInternalServerError), http.StatusInternalServerError)
-			return
-		}
-
-		if _, err := fmt.Fprint(writer, string(finalJson)); err != nil {
-			log.Print(err)
-			http.Error(writer, http.StatusText(http.StatusInternalServerError), http.StatusInternalServerError)
-			return
-		}
-	}
-}
-
-func parseSearchPhrase(request *http.Request) ([]string, error, int) {
-
-	searchPhrase := request.FormValue("search")
-
-	rawUserInput := strings.ToLower(searchPhrase)
-
-	parsedUserInput := strings.Split(rawUserInput, " ")
-
-	cleanedUserInput := make([]string, 0)
-	for i := range parsedUserInput {
-		w, err := util.CleanUserData(parsedUserInput[i])
-		if err != nil {
-			log.Print(err)
-			return nil, err, http.StatusBadRequest
-		}
-		if w != "" {
-			cleanedUserInput = append(cleanedUserInput, w)
-		}
-	}
-	return cleanedUserInput, nil, -1
-}
-
-func answerFormation(index *Index, cleanedUserInput []string) ([]*searchResponse, error, int) {
-
-	ans, err := index.buildSearchIndex(cleanedUserInput)
-	if err != nil {
-		log.Print(err)
-		return nil, err, http.StatusInternalServerError
-	}
-
-	var resp []*searchResponse
-	for s := range ans {
-		resp = append(resp, &searchResponse{
-			Filename:    s,
-			WordCounter: ans[s],
-		})
-		log.Printf("filename: %s, words encountered : %d", s, ans[s])
-	}
-	return resp, nil, -1
-}
-
-func unmarshalFile(filename string) (*Index, error) {
-	var m *Index
-	content, err := ioutil.ReadFile(filename)
-	if err != nil {
-		log.Print(err)
-		return nil, err
-	}
-
-	if json.Unmarshal(content, &m) != nil {
-		log.Print(err)
-		return nil, err
-	}
-	return m, nil
 }
